*增加了LocationString直接通过字符串构造方法 *添加FoundUnfinishedSearchDto对接类，用于和前端搜索对接 *修改HomeItemDataController请求参数类型 *将FoundUnfinishedDto删除，换为更统一的ItemDto数据传输类

+++ src/main/resources/schema.sql
@@ -102,7 +102,7 @@ CREATE TABLE IF NOT EXISTS colors (
 );
 
 CREATE TABLE IF NOT EXISTS item_colors (
-     item_id INT,
+     item_id BIGINT,
      color_id INT,
      PRIMARY KEY(item_id, color_id),
      FOREIGN KEY (color_id) REFERENCES colors(id)
